Input parameter "sweeps-per-measurement" takes integer values.

+++ include/dca/phys/dca_step/cluster_solver/ctaux/ctaux_walker.hpp
@@ -376,7 +376,7 @@ void CtauxWalker<device_t, parameters_type, MOMS_type>::initialize() {
 
 template <dca::linalg::DeviceType device_t, class parameters_type, class MOMS_type>
 void CtauxWalker<device_t, parameters_type, MOMS_type>::do_sweep() {
-  double factor = 1.;
+  int factor = 1;
   if (thermalized)
     factor = parameters.get_sweeps_per_measurement();
